requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiohttp-retry==2.9.1
aiosignal==1.4.0
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.13.0
appdirs==1.4.4
arrow==1.4.0
asyncssh==2.22.0
atpublic==7.0.0
attrs==26.1.0
billiard==4.2.4
binaryornot==0.6.0
celery==5.6.3
certifi==2026.4.22
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.3
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
configobj==5.0.9
cookiecutter==2.7.1
cryptography==46.0.7
dictdiffer==0.9.0
diskcache==5.6.3
distro==1.9.0
dnspython==2.8.0
dpath==2.2.0
dulwich==1.2.0
dvc==3.67.1
dvc-data==3.18.3
dvc-http==2.32.0
dvc-objects==5.2.0
dvc-render==1.0.2
dvc-studio-client==0.22.0
dvc-task==0.40.2
email-validator==2.3.0
entrypoints==0.4
fastapi==0.136.0
fastapi-cli==0.0.24
fastapi-cloud-cli==0.17.0
fastar==0.11.0
filelock==3.29.0
flatten-dict==0.4.2
flufl.lock==9.0.0
frozenlist==1.8.0
fsspec==2026.3.0
funcy==2.0
gitdb==4.0.12
GitPython==3.1.47
grandalf==0.8
gto==1.9.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
hydra-core==1.3.2
idna==3.12
iterative-telemetry==0.0.10
Jinja2==3.1.6
joblib==1.5.3
kombu==5.6.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
multidict==6.7.1
networkx==3.6.1
numpy==2.4.4
omegaconf==2.3.0
orjson==3.11.8
packaging==26.1
pandas==3.0.2
pathspec==1.0.4
platformdirs==4.9.6
prompt_toolkit==3.0.52
propcache==0.4.1
psutil==7.2.2
pycparser==3.0
pydantic==2.13.3
pydantic-extra-types==2.11.1
pydantic-settings==2.14.0
pydantic_core==2.46.3
pydot==4.0.1
pygit2==1.19.2
Pygments==2.20.0
pygtrie==2.5.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.26
python-slugify==8.0.4
PyYAML==6.0.3
requests==2.33.1
rich==15.0.0
rich-toolkit==0.19.7
rignore==0.7.6
ruamel.yaml==0.19.1
scikit-learn==1.8.0
scipy==1.17.1
scmrepo==3.6.2
semver==3.0.4
sentry-sdk==2.58.0
setuptools==82.0.1
shellingham==1.5.4
shortuuid==1.0.13
shtab==1.8.0
six==1.17.0
smmap==5.0.3
sqltrie==0.11.2
starlette==1.0.0
tabulate==0.10.0
text-unidecode==1.3
threadpoolctl==3.6.0
tomlkit==0.14.0
tqdm==4.67.3
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
tzlocal==5.3.1
urllib3==2.6.3
uvicorn==0.45.0
uvloop==0.22.1
vine==5.1.0
voluptuous==0.16.0
watchfiles==1.1.1
wcwidth==0.6.0
websockets==16.0
yarl==1.23.0
zc.lockfile==4.0
